Allow host to configure pattern generator.
This enables using the write-reg32 API to configure which
pattern generator to enable, as well as to provide a custom
pattern generator with up to 4 patterns allowed.

Please note that enabling these patterns, especially a custom
pattern, may cause the NIC to become an intefering source and
to run afoul of RF regulation bodies.  These features should only
be used by someone with full knowledge of what this can do,
preferably in a lab environment and/or in an isolation chamber.

Signed-off-by: Ben Greear <greearb@candelatech.com>